requirements.txt
setuptools-scm>=3.3.3
numpy>=1.15.0
scipy>=1.0.0
fastcluster
pandas>=1.0.1
matplotlib
seaborn>=0.10.0
scikit-learn>=0.19.1
statsmodels>=0.8.0
patsy>=0.4.1
requests>=2.21.0
umap-learn

fancyimpute

imc
